Tracking progress uploading track

+++ EkofyApp.Domain/Entities/Artist.cs
@@ -24,7 +24,7 @@ public sealed class Artist : TimeStamped, IEntityCustom
 
     public string? Biography { get; set; } // Biography of the artist, e.g., "John Doe is a singer-songwriter from..."
     public long FollowerCount { get; set; } = default; // Number of followers the artist has
-    public long Popularity { get; set; } = default; // Popularity score of the artist
+    public decimal Popularity { get; set; } = default; // Popularity score of the artist
     public string? AvatarImage { get; set; } // URL to the artist's avatar image
     public string? BannerImage { get; set; } // URL to the artist's banner image
